lambda-build-jwt/requirements.txt
# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Snowflake connector with JWT support (compatible version)
snowflake-connector-python>=1.9.1

# Cryptography for JWT authentication (use current working version)
cryptography

# HTTP requests
requests>=2.31.0

# JSON handling
orjson>=3.9.10

# Date/time handling
python-dateutil>=2.8.2

# Environment variable handling
python-dotenv>=1.0.0
